requirements.txt
fastapi==0.115.5
tensorflow==2.12.0
keras==2.12.0
numpy==1.23.5
Pillow==11.0.0
uvicorn
python-multipart
